fix advancedSearch
