+++ app/frontend/src/classes/AdvancedSearch/AdvancedSearchConditionRestorer.ts
@@ -1,34 +1,24 @@
-import { ADVANCED_CONDITIONS, API_URL } from '../../global';
-import {
-  CONDITION_TYPE,
-  type ConditionTypeValue,
-  type FrequencyDataset,
-} from '../../definition';
-import {
-  PREDICTIONS,
-  type PredictionKey,
-} from '../../components/ConditionPathogenicityPredictionSearch/PredictionDatasets';
-import { axios } from '../../utils/cachedAxios';
+import { CONDITION_TYPE, type ConditionTypeValue } from '../../definition';
 import type { ConditionGroupView } from '../Condition/ConditionGroupView';
-import type {
-  RestoredFrequencyMode,
-  RestoredConditionValue,
-  RestoredPredictionValue,
-} from '../Condition/ConditionItemView';
-import type {
-  Inequality,
-  LogicalOperator,
-  Relation,
-  SignificanceSource,
-} from '../../types';
-
-type QueryObject = Record<string, unknown>;
-
-type RestoredItem = Readonly<{
-  conditionType: ConditionTypeValue;
-  relation?: Relation;
-  values: RestoredConditionValue[];
-}>;
+import type { LogicalOperator } from '../../types';
+import {
+  toMergedFrequencyItem,
+  restoreFrequencyItem,
+} from './AdvancedSearchFrequencyRestorer';
+import { restoreGeneItem } from './AdvancedSearchGeneRestorer';
+import {
+  restoreSignificanceItem,
+  restorePredictionItem,
+} from './AdvancedSearchSignificanceRestorer';
+import {
+  isQueryObject,
+  makeValue,
+  getRangeStart,
+  getRangeEnd,
+  findConditionLabel,
+  type QueryObject,
+  type RestoredItem,
+} from './AdvancedSearchRestorerUtils';
 
 /**
  * URLから復元したAdvanced Search条件を、BuilderのView構造へ戻す。
@@ -47,7 +37,7 @@ export async function restoreAdvancedSearchCondition(
   await appendQueryToGroup(rootGroup, query, true);
 }
 
-// query fragmentを再帰的にたどり、論理グループまたは条件行として追加する。
+/** query fragmentを再帰的にたどり、論理グループまたは条件行として追加する。 */
 async function appendQueryToGroup(
   parentGroup: ConditionGroupView,
   query: unknown,
@@ -57,6 +47,7 @@ async function appendQueryToGroup(
 
   const logical = getLogicalQuery(query);
   if (logical) {
+    // dataset/genotypeの複数選択はorで展開されているため、先にまとめを試みる。
     const mergedItem = toMergedFrequencyItem(logical);
     if (mergedItem) {
       await appendRestoredItem(parentGroup, mergedItem);
@@ -69,6 +60,7 @@ async function appendQueryToGroup(
       : parentGroup.addEmptyConditionGroup(logical.operator);
     targetGroup.logicalOperator = logical.operator;
 
+    // autoUngroupを一時停止しないと、子の追加ごとに不要なungroup処理が走る。
     const resumeAutoUngroup = useCurrentGroup
       ? null
       : targetGroup.suspendAutoUngroup();
@@ -88,11 +80,11 @@ async function appendQueryToGroup(
   await appendRestoredItem(parentGroup, item);
 }
 
+/** ConditionItemViewを通常生成した後に値だけを注入し、編集モードを閉じた状態に戻す。 */
 async function appendRestoredItem(
   parentGroup: ConditionGroupView,
   item: RestoredItem
 ): Promise<void> {
-  // ConditionItemViewの通常生成後に値だけを注入し、編集モードを閉じた状態に戻す。
   const itemView = parentGroup.addNewConditionItem(item.conditionType);
   await itemView.hydrateFromRestoredQuery({
     relation: item.relation,
@@ -100,7 +92,7 @@ async function appendRestoredItem(
   });
 }
 
-// { and: [...] } / { or: [...] } だけを論理グループとして扱う。
+/** { and: [...] } / { or: [...] } だけを論理グループとして扱う。 */
 function getLogicalQuery(
   query: QueryObject
 ): { operator: LogicalOperator; children: unknown[] } | null {
@@ -113,7 +105,7 @@ function getLogicalQuery(
   return { operator, children };
 }
 
-// query leafのキーから、復元すべき条件種別と表示値へ変換する。
+/** query leafのキーから、復元すべき条件種別と表示値へ変換するディスパッチャ。 */
 async function toRestoredItem(
   query: QueryObject
 ): Promise<RestoredItem | null> {
@@ -143,7 +135,7 @@ async function toRestoredItem(
   ] as const) {
     const leaf = query[conditionType];
     if (isQueryObject(leaf)) {
-      return restoreTermItem(conditionType, leaf, false);
+      return restoreTermItem(conditionType, leaf);
     }
   }
 
@@ -151,227 +143,10 @@ async function toRestoredItem(
     return restoreSignificanceItem(query.significance);
   }
 
-  const prediction = restorePredictionItem(query);
-  if (prediction) return prediction;
-
-  return null;
+  return restorePredictionItem(query);
 }
 
-// dataset/genotypeの複数選択はquery上では { or: [frequency, ...] } になるため、UI復元時は1条件行に畳み戻す。
-function toMergedFrequencyItem(logical: {
-  operator: LogicalOperator;
-  children: unknown[];
-}): RestoredItem | null {
-  if (logical.operator !== 'or') return null;
-
-  const items = logical.children
-    .map((child) =>
-      isQueryObject(child) && isQueryObject(child.frequency)
-        ? restoreFrequencyItem(child.frequency)
-        : null
-    )
-    .filter((item): item is RestoredItem => item !== null);
-
-  if (items.length !== logical.children.length || items.length <= 1) {
-    return null;
-  }
-
-  const conditionType = items[0].conditionType;
-  const canMerge = items.every(
-    (item) =>
-      item.conditionType === conditionType &&
-      item.values.length > 0 &&
-      item.values.every((value) => value.frequency)
-  );
-  if (!canMerge) return null;
-
-  return {
-    conditionType,
-    values: items.flatMap((item) => item.values),
-  };
-}
-
-// Gene symbolはURL上では数値IDだけを持つため、表示用symbolをAPIから引き直す。
-async function restoreGeneItem(
-  gene: QueryObject
-): Promise<RestoredItem | null> {
-  const terms = gene.terms;
-  if (!Array.isArray(terms)) return null;
-
-  const relation = gene.relation === 'ne' ? 'ne' : 'eq';
-  const values = await Promise.all(
-    terms.map(async (term) => {
-      const value = String(term);
-      const label =
-        getGeneLabelFromQuery(gene.labels, value) ??
-        (await findGeneSymbolLabel(value)) ??
-        value;
-      return makeValue(value, label);
-    })
-  );
-
-  return {
-    conditionType: CONDITION_TYPE.gene_symbol,
-    relation,
-    values,
-  };
-}
-
-function getGeneLabelFromQuery(labels: unknown, geneId: string): string | null {
-  if (!isQueryObject(labels)) return null;
-  return getString(labels[geneId]);
-}
-
-// Pathogenicity predictionはquery key(alphamissense/sift/polyphen)から共通の条件行へ戻す。
-function restorePredictionItem(query: QueryObject): RestoredItem | null {
-  const predictionKey = getPredictionKey(query);
-  if (!predictionKey) return null;
-
-  const leaf = query[predictionKey];
-  if (!isQueryObject(leaf)) return null;
-
-  const prediction = toRestoredPredictionValue(predictionKey, leaf.score);
-  if (!prediction) return null;
-
-  return {
-    conditionType: CONDITION_TYPE.pathogenicity_prediction,
-    values: [
-      {
-        value: predictionKey,
-        label: PREDICTIONS[predictionKey].label,
-        prediction,
-      },
-    ],
-  };
-}
-
-function getPredictionKey(query: QueryObject): PredictionKey | null {
-  return (Object.keys(PREDICTIONS) as PredictionKey[]).find((key) =>
-    isQueryObject(query[key])
-  ) ?? null;
-}
-
-function toRestoredPredictionValue(
-  dataset: PredictionKey,
-  score: unknown
-): RestoredPredictionValue | null {
-  if (isQueryObject(score)) {
-    return {
-      dataset,
-      values: [getRangeStart(score) ?? 0, getRangeEnd(score) ?? 1],
-      inequalitySigns: [getLeftInequality(score), getRightInequality(score)],
-      includeUnassigned: false,
-      includeUnknown: false,
-    };
-  }
-
-  if (Array.isArray(score)) {
-    return {
-      dataset,
-      values: [0, 0],
-      inequalitySigns: ['gt', 'lt'],
-      includeUnassigned: score.includes('unassigned'),
-      includeUnknown: score.includes('unknown'),
-    };
-  }
-
-  return null;
-}
-
-function getLeftInequality(
-  score: QueryObject
-): Extract<Inequality, 'gte' | 'gt'> {
-  return score.gt === undefined ? 'gte' : 'gt';
-}
-
-function getRightInequality(
-  score: QueryObject
-): Extract<Inequality, 'lte' | 'lt'> {
-  return score.lt === undefined ? 'lte' : 'lt';
-}
-
-async function findGeneSymbolLabel(geneId: string): Promise<string | null> {
-  const url = new URL(`${API_URL}/api/search/${CONDITION_TYPE.gene_symbol}`);
-  url.searchParams.set('term', geneId);
-
-  try {
-    const { data } = await axios.get(url.toString());
-    const suggestions = Array.isArray(data) ? data : [];
-    const matched = suggestions.find(
-      (suggestion) =>
-        isQueryObject(suggestion) && String(suggestion.id) === geneId
-    );
-    if (!isQueryObject(matched)) return null;
-
-    return getString(matched.symbol ?? matched.label ?? matched.name);
-  } catch (_error) {
-    return null;
-  }
-}
-
-// dataset/genotype条件は、外側の値表示と内側のfrequency-count-value-viewの状態を両方復元する。
-function restoreFrequencyItem(frequency: QueryObject): RestoredItem | null {
-  const dataset = frequency.dataset;
-  if (!isQueryObject(dataset) || typeof dataset.name !== 'string') return null;
-
-  const datasetName = dataset.name as FrequencyDataset;
-  const genotype = frequency.genotype;
-  const isGenotype = isQueryObject(genotype);
-  const conditionType = isGenotype
-    ? CONDITION_TYPE.genotype
-    : CONDITION_TYPE.dataset;
-
-  const mode = getFrequencyMode(frequency, genotype);
-  const range = getFrequencyRange(frequency, genotype);
-  if (!mode || !isQueryObject(range)) return null;
-
-  return {
-    conditionType,
-    values: [
-      {
-        value: datasetName,
-        label: findConditionLabel(conditionType, datasetName),
-        frequency: {
-          conditionType,
-          mode,
-          from: getRangeStart(range),
-          to: getRangeEnd(range),
-          invert: false,
-          filtered: frequency.filtered === true,
-        },
-      },
-    ],
-  };
-}
-
-// genotypeは genotype.key、datasetは frequency/count のどちらを持つかでUIモードを決める。
-function getFrequencyMode(
-  frequency: QueryObject,
-  genotype: unknown
-): RestoredFrequencyMode | null {
-  if (isQueryObject(genotype) && typeof genotype.key === 'string') {
-    return toSupportedFrequencyMode(genotype.key);
-  }
-  if (isQueryObject(frequency.frequency)) return 'frequency';
-  if (isQueryObject(frequency.count)) return 'count';
-  return null;
-}
-
-// UI側が対応しているgenotypeモードだけを復元対象にする。
-function toSupportedFrequencyMode(value: string): RestoredFrequencyMode | null {
-  return value === 'aac' || value === 'arc' || value === 'hac' ? value : null;
-}
-
-// query内でrangeが入っている位置はdataset条件とgenotype条件で異なる。
-function getFrequencyRange(
-  frequency: QueryObject,
-  genotype: unknown
-): unknown {
-  if (isQueryObject(genotype)) return genotype.count;
-  return frequency.frequency ?? frequency.count;
-}
-
-// API用のlocation queryを、Location editorが扱う "chr:start-end" 形式へ戻す。
+/** API用のlocation queryを、Location editorが扱う "chr:start-end" 形式へ戻す。 */
 function restoreLocationItem(location: QueryObject): RestoredItem | null {
   const chromosome = location.chromosome;
   if (typeof chromosome !== 'string') return null;
@@ -381,9 +156,7 @@ function restoreLocationItem(location: QueryObject): RestoredItem | null {
     typeof position === 'number'
       ? `${chromosome}:${position}`
       : isQueryObject(position)
-      ? `${chromosome}:${getRangeStart(position) ?? ''}-${
-          getRangeEnd(position) ?? ''
-        }`
+      ? `${chromosome}:${getRangeStart(position) ?? ''}-${getRangeEnd(position) ?? ''}`
       : null;
 
   if (!value) return null;
@@ -394,49 +167,10 @@ function restoreLocationItem(location: QueryObject): RestoredItem | null {
   };
 }
 
-// Clinical significanceはsourceごとに表示ラベル(MGeND/ClinVar)とqueryを分ける必要がある。
-function restoreSignificanceItem(
-  significance: QueryObject
-): RestoredItem | null {
-  const terms = significance.terms;
-  if (!Array.isArray(terms)) return null;
-
-  const sources = getSignificanceSources(significance.source);
-  if (sources.length === 0) return null;
-
-  const relation = significance.relation === 'ne' ? 'ne' : 'eq';
-  return {
-    conditionType: CONDITION_TYPE.significance,
-    relation,
-    values: sources.flatMap((source) =>
-      terms.map((term) => {
-        const value = String(term);
-        return {
-          ...makeValue(
-            value,
-            findConditionLabel(CONDITION_TYPE.significance, value)
-          ),
-          source,
-        };
-      })
-    ),
-  };
-}
-
-function getSignificanceSources(source: unknown): SignificanceSource[] {
-  if (!Array.isArray(source)) return [];
-
-  return source.filter(
-    (item): item is SignificanceSource =>
-      item === 'mgend' || item === 'clinvar'
-  );
-}
-
-// relationとtermsを持つ標準的な条件を、共通の表示値配列へ戻す。
+/** relationとtermsを持つ標準的な条件を、共通の表示値配列へ戻す。 */
 function restoreTermItem(
   conditionType: ConditionTypeValue,
-  leaf: QueryObject,
-  numericTerms: boolean
+  leaf: QueryObject
 ): RestoredItem | null {
   const terms = leaf.terms;
   if (!Array.isArray(terms)) return null;
@@ -446,81 +180,8 @@ function restoreTermItem(
     conditionType,
     relation,
     values: terms.map((term) => {
-      const value = numericTerms ? String(Number(term)) : String(term);
+      const value = String(term);
       return makeValue(value, findConditionLabel(conditionType, value));
     }),
   };
-}
-
-// ラベルが見つからない条件でもUI表示できるよう、valueをlabelの代替にする。
-function makeValue(value: string, label: string = value): RestoredConditionValue {
-  return { value, label };
-}
-
-// gt/gteの違いは現状の表示UIでは区別しないため、開始値としてまとめて扱う。
-function getRangeStart(range: QueryObject): number | null {
-  return getNumber(range.gte ?? range.gt);
-}
-
-// lt/lteの違いは現状の表示UIでは区別しないため、終了値としてまとめて扱う。
-function getRangeEnd(range: QueryObject): number | null {
-  return getNumber(range.lte ?? range.lt);
-}
-
-// URL由来の値は信用せず、有限なnumberだけをrange値として採用する。
-function getNumber(value: unknown): number | null {
-  return typeof value === 'number' && Number.isFinite(value) ? value : null;
-}
-
-function getString(value: unknown): string | null {
-  return typeof value === 'string' && value.length > 0 ? value : null;
-}
-
-// 検索条件マスタから表示ラベルを探し、URL復元後も通常操作時に近い見た目へ戻す。
-function findConditionLabel(
-  conditionType: ConditionTypeValue,
-  value: string
-): string {
-  const condition = ADVANCED_CONDITIONS[conditionType];
-  if (!condition || typeof condition !== 'object' || !('values' in condition)) {
-    return value;
-  }
-
-  return findLabelInValues(condition.values, value) ?? value;
-}
-
-// datasetのような階層値と、significanceのようなsource別値の両方を再帰的に探す。
-function findLabelInValues(values: unknown, targetValue: string): string | null {
-  if (Array.isArray(values)) {
-    for (const item of values) {
-      const label = findLabelInValueItem(item, targetValue);
-      if (label) return label;
-    }
-    return null;
-  }
-
-  if (isQueryObject(values)) {
-    for (const value of Object.values(values)) {
-      const label = findLabelInValues(value, targetValue);
-      if (label) return label;
-    }
-  }
-
-  return null;
-}
-
-// 1つのマスタ項目を調べ、子要素があればさらに下へ潜る。
-function findLabelInValueItem(item: unknown, targetValue: string): string | null {
-  if (!isQueryObject(item)) return null;
-
-  if (item.value === targetValue && typeof item.label === 'string') {
-    return item.label;
-  }
-
-  return findLabelInValues(item.children, targetValue);
-}
-
-// URL由来のunknownを安全に扱うため、配列ではないプレーンなobjectだけに絞る。
-function isQueryObject(value: unknown): value is QueryObject {
-  return typeof value === 'object' && value !== null && !Array.isArray(value);
 }
